Modularize chat and UI
